requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.12
python-dotenv==1.2.2